server/requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
argcomplete==3.3.0
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
commitizen==3.24.0
dataclasses-json==0.6.4
decli==0.6.1
deprecation==2.1.0
distro==1.9.0
faiss-cpu==1.8.0
fastapi==0.109.2
frozenlist==1.4.1
gotrue==2.1.0
h11==0.14.0
httpcore==1.0.3
httpx==0.25.2
idna==3.6
importlib_metadata==7.1.0
Jinja2==3.1.3
jsonpatch==1.33
jsonpointer==2.4
langchain==0.1.7
langchain-community==0.0.20
langchain-core==0.1.23
langchain-openai==0.0.6
langsmith==0.0.87
MarkupSafe==2.1.5
marshmallow==3.20.2
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
openai==1.12.0
packaging==23.2
postgrest==0.15.0
prompt-toolkit==3.0.36
psycopg2-binary==2.9.9
pydantic==2.6.1
pydantic_core==2.16.2
python-dateutil==2.8.2
python-dotenv==1.0.1
PyYAML==6.0.1
questionary==2.0.1
realtime==1.0.2
regex==2023.12.25
requests==2.31.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.27
starlette==0.36.3
storage3==0.7.0
StrEnum==0.4.15
supabase==2.3.5
supabase-py-async==2.5.6
supafunc==0.4.0
tenacity==8.2.3
termcolor==2.4.0
tiktoken==0.6.0
tomlkit==0.12.4
tqdm==4.66.2
typing-inspect==0.9.0
typing_extensions==4.9.0
urllib3==2.2.0
uvicorn==0.27.1
wcwidth==0.2.13
websockets==11.0.3
yarl==1.9.4
zipp==3.18.1
